extract DB stuff to separate projec. Add another controller for reviews
