+++ Database/Data/Review.cs
@@ -5,7 +5,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace HotelWebsite.Data
+namespace Database.Data
 {
     public class Review
     {
