+++ HotelWebsite.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29318.209
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HotelWebsite", "HotelWebsite\HotelWebsite.csproj", "{5CC237D9-66A2-48E6-B875-68B2968D0B56}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HotelWebsite", "HotelWebsite\HotelWebsite.csproj", "{5CC237D9-66A2-48E6-B875-68B2968D0B56}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Database", "Database\Database.csproj", "{1025F9CF-6DAF-4034-BD19-4D07B542FE05}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{5CC237D9-66A2-48E6-B875-68B2968D0B56}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5CC237D9-66A2-48E6-B875-68B2968D0B56}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5CC237D9-66A2-48E6-B875-68B2968D0B56}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1025F9CF-6DAF-4034-BD19-4D07B542FE05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1025F9CF-6DAF-4034-BD19-4D07B542FE05}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1025F9CF-6DAF-4034-BD19-4D07B542FE05}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1025F9CF-6DAF-4034-BD19-4D07B542FE05}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
