+++ HotelWebsite/Areas/Identity/IdentityHostingStartup.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using HotelWebsite.Data;
+using Database.Data;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.UI;
