+++ HotelWebsite/Models/CreateReviewViewModel.cs
@@ -1,4 +1,4 @@
-﻿using HotelWebsite.Data;
+﻿using Database.Data;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
