+++ HotelWebsite/Startup.cs
@@ -1,4 +1,4 @@
-using HotelWebsite.Data;
+using Database.Data;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Identity;
