+++ HotelWebsite/Views/Offer/Offer.cshtml
@@ -38,7 +38,7 @@
                 </div>
             </div>
             <div class="product-review-links">
-                <a asp-action="CreateReview" asp-controller="Offer" asp-route-id="@Model.ID">Leave a Review</a>
+                <a asp-action="CreateReview" asp-controller="Review" asp-route-id="@Model.ID">Leave a Review</a>
             </div>
         </div>
         <div class="prices"><div class="product-price"><span>$@Model.Price</span></div></div>
